Draft Architecture
This model takes three input matrices: the node data, an adjacency matrix, and a behavioural data matrix. GNN and LSTM layers are used in the main network. The behavioural data is embedded using a separate MLP and then combined with the main network data rather than being directly incorporated into the node matrix.